Fix clang-tidy warnings

+++ pdns/webserver.cc
@@ -587,7 +587,7 @@ WebServer::WebServer(string listenaddress, int port) :
   d_listenaddress(std::move(listenaddress)),
   d_port(port),
   d_server(nullptr),
-  d_maxbodysize(2 * 1024 * 1024)
+  d_maxbodysize(static_cast<ssize_t>(2 * 1024 * 1024))
 
 {
     YaHTTP::Router::Map("OPTIONS", "/<*url>", [](YaHTTP::Request *req, YaHTTP::Response *resp) {
